requirements.txt
asgiref==3.2.7
astroid==2.4.2
beautifulsoup4==4.9.0
bios==0.1.2
bson==0.5.8
captcha==0.3
certifi==2020.4.5.1
chardet==3.0.4
codegen==1.0
colorama==0.4.3
cycler==0.10.0
dataclasses==0.6
Django==2.2.10
django-crispy-forms==1.9.2
django-dirtyfields==1.4
django-extensions==2.2.7
django-filter==2.3.0
django-plugins==0.3.0
djangorestframework==3.11.0
djongo==1.3.2
docker==4.4.0
geographiclib==1.50
geopy==2.0.0
groups==0.1
gunicorn==20.0.4
idna==2.9
isort==5.7.0
jsonparser==1.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
Markdown==3.2.2
matplotlib==3.3.3
mccabe==0.6.1
mysql-connector-python==8.0.22
mysqlclient==2.0.1
numpy==1.19.3
oyaml==1.0
pandas==1.1.4
patterns==0.3
Pillow==7.1.2
PyJWT==2.0.0
pylint==2.6.0
pymongo==3.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
PyYAML==5.4.1
qrcode==6.1
requests==2.23.0
sinchsms==1.0.4
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
toml==0.10.2
twilio==6.50.1
urllib3==1.25.9
websocket-client==0.57.0
Werkzeug==1.0.0
wrapt==1.12.1
xlrd==2.0.1
